requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
transformers
peft
accelerate
pillow
qwen-vl-utils